DjangoBICAP/DjangoBICAP/requirements.txt
django >= 2.1.6, < 3.0
djangorestframework>=3.9.0